@@ -7,9 +7,12 @@
   },
   "dependencies": {
     "bootstrap-without-jquery": "^1.0.4",
-    "react": "^15.2.1",
-    "react-dom": "^15.2.1",
-    "react-router": "2.7.0"
+    "material-ui": "^0.15.4",
+    "react-tap-event-plugin": "^1.0.0",
+    "react": "^15.3.2",
+    "react-dom": "^15.3.2",
+    "react-addons-pure-render-mixin": "~15.3.2",
+    "react-router": "2.8.1"
   },
   "scripts": {
     "start": "react-scripts start",
